Add solution for problem `A - Not Overflow` Link: https://atcoder.jp/contests/abc237/tasks/abc237_a Tags: AtCoder, *100 Ref: https://atcoder.jp/contests/abc237/submissions/28988703

+++ atcoder/abc237/A.cpp
@@ -4,7 +4,7 @@ using namespace std;
 int main(){
   long long n;
   cin >> n;
-  long long two31 = powl(2,31);
+  long long two31 = pow(2,31);
   if(n < -two31 || n >= two31){
     cout << "No\n";
   } else cout << "Yes\n";
